remove previous Stardust identity resolver, minor cleanup
