+++ client/src/app/components/identity/IdentityStardustResolver.tsx
@@ -1,32 +1,25 @@
+import * as identity from "@iota/identity-wasm/web";
+import { DomainLinkageConfiguration, EdDSAJwsVerifier, JwtCredentialValidationOptions, JwtDomainLinkageValidator, LinkedDomainService } from "@iota/identity-wasm/web";
 import React, { Fragment, ReactNode } from "react";
-import { RouteComponentProps } from "react-router-dom";
+import IdentityDomainResolver from "./domains/IdentityDomainResolver";
 import { IdentityStardustResolverProps } from "./IdentityStardustResolverProps";
 import { IdentityStardustResolverState } from "./IdentityStardustResolverState";
-import welcomeMessage from "../../../assets/modals/identity-resolver/welcome.json";
 import { ServiceFactory } from "../../../factories/serviceFactory";
-import { IdentityDiffStorageService } from "../../../services/identityDiffStorageService";
 import { IdentityService } from "../../../services/identityService";
 import AsyncComponent from "../AsyncComponent";
 import "./IdentityStardustResolver.scss";
 import CopyButton from "../CopyButton";
 import JsonViewer from "../JsonViewer";
-import Modal from "../Modal";
 import Spinner from "../Spinner";
 
-class IdentityStardustResolver extends AsyncComponent<
-    RouteComponentProps<IdentityStardustResolverProps>,
+class IdentityStardustResolver extends AsyncComponent<IdentityStardustResolverProps,
     IdentityStardustResolverState
 > {
-    /**
-     * Timer to check to state update.
-     */
-    private _timerId?: NodeJS.Timer;
-
-    constructor(props: RouteComponentProps<IdentityStardustResolverProps>) {
+    constructor(props: IdentityStardustResolverProps) {
         super(props);
         this.state = {
-            did: props.match.params.did,
-            aliasId: getAliasId(props.match.params.did ?? ""),
+            did: props.did,
+            aliasId: getAliasId(props.did ?? ""),
             errorMessage: ""
         };
     }
@@ -40,7 +33,7 @@ class IdentityStardustResolver extends AsyncComponent<
 
         const resolvedIdentity = await ServiceFactory.get<IdentityService>("identity").resolveIdentityStardust(
             this.state.did,
-            this.props.match.params.network
+            this.props.network
         );
 
         if (resolvedIdentity.error) {
@@ -53,176 +46,195 @@ class IdentityStardustResolver extends AsyncComponent<
         this.setState({
             resolvedIdentity
         });
-    }
 
-    /**
-     * The component will unmount so update flag.
-     */
-    public componentWillUnmount(): void {
-        super.componentWillUnmount();
-        if (this._timerId) {
-            clearTimeout(this._timerId);
-            this._timerId = undefined;
+        await identity.init("/identity_wasm_bg.wasm");
+
+        const didDocument = identity.IotaDocument.fromJSON(resolvedIdentity.document);
+        // Get the Linked Domain Services from the DID Document.
+        const linkedDomainServices = didDocument
+            .service()
+            .filter(service => LinkedDomainService.isValid(service))
+            .map(service => LinkedDomainService.fromService(service));
+
+        const verifiedDomains = new Map<string, Promise<void>>();
+
+        for (const entry of linkedDomainServices) {
+            for (const domain of entry.domains()) {
+                verifiedDomains.set(domain, new Promise(async (resolve, reject) => {
+                    // Note that according to the specs, the DID Configuration resource must exist
+                    // at the origin's root, Well-Known Resource directory.
+                    const configurationUrl = new URL("/.well-known/did-configuration.php", domain); // TODO: .json
+
+                    let fetchedConfigurationResource;
+
+                    try {
+                        fetchedConfigurationResource = await fetch(configurationUrl);
+                    } catch (err) {
+                        console.log(err);
+                        reject(new Error(`could not fetch configuration from ${domain}`));
+                        return;
+                    }
+
+                    let parsedConfigurationResource;
+
+                    try {
+                        const jsonResponse = await fetchedConfigurationResource.json();
+                        parsedConfigurationResource = DomainLinkageConfiguration.fromJSON(jsonResponse);
+                    } catch (err) {
+                        console.log(err);
+                        reject(new Error(`could not parse configuration from domain ${domain}`));
+                        return;
+                    }
+
+                    try {
+                        new JwtDomainLinkageValidator(new EdDSAJwsVerifier()).validateLinkage(
+                            didDocument,
+                            parsedConfigurationResource,
+                            domain,
+                            new JwtCredentialValidationOptions()
+                        );
+                        resolve();
+                    } catch (err) {
+                        reject(err);
+                    }
+                }));
+            }
         }
-        IdentityDiffStorageService.instance.clearAll();
-    }
 
+        this.setState({
+            verifiedDomains
+        });
+    }
 
     /**
      * Render the component.
      * @returns The node to render.
      */
     public render(): ReactNode {
-        const network = this.props.match.params.network;
+        const network = this.props.network;
         const governorAddress = this.state.resolvedIdentity?.document?.meta.governorAddress;
         const stateControllerAddress = this.state.resolvedIdentity?.document?.meta.stateControllerAddress;
         return (
             <div>
-                <div className="row space-between wrap">
-                    <div className="row middle">
-                        <h1>
-                            Decentralized Identifier
-                        </h1>
-                        <Modal icon="info" data={welcomeMessage} />
-                    </div>
+
+
+                <div className="label">DID</div>
+                <div className="row middle value code highlight margin-b-s">
+                    <div className="margin-r-t">{this.props.did}</div>
+                    <CopyButton copy={this.props.did} />
                 </div>
-                <div className="section">
-                    <div className="section--data">
-                        <div className="label">DID</div>
-                        <div className="row middle value code highlight margin-b-s">
-                            <div className="margin-r-t">{this.props.match.params.did}</div>
-                            <CopyButton copy={this.props.match.params.did} />
-                        </div>
 
-                        {this.state.resolvedIdentity?.document &&
-                            !this.state.errorMessage && (
-                                <Fragment>
-                                    <div className="margin-b-s">
-                                        <div className="label">Alias ID</div>
-                                        <div className="value code row middle">
-                                            <div className="margin-r-t">
-                                                <a onClick={() => {
+                {this.state.resolvedIdentity?.document &&
+                        !this.state.errorMessage && (
+                            <Fragment>
+                                <div className="margin-b-s">
+                                    <div className="label">Governor</div>
+                                    <div className="value code row middle">
+                                        <div className="margin-r-t">
+                                            <a onClick={() => {
+                                                window.location.href =
                                                     // eslint-disable-next-line max-len
-                                                    window.location.href = `/${network}/search/${this.state.aliasId}`;
-                                                }}
-                                                >
-                                                    {this.state.aliasId}
-                                                </a>
-                                            </div>
-                                            <CopyButton
-                                                copy={this.state.aliasId}
-                                            />
+                                                    `/${network}/search/${governorAddress}`;
+                                            }}
+                                            >
+                                                {governorAddress}
+                                            </a>
                                         </div>
+                                        <CopyButton
+                                            copy={this.state.resolvedIdentity.document.meta.governorAddress}
+                                        />
                                     </div>
-                                    <div className="margin-b-s">
-                                        <div className="label">Governor</div>
-                                        <div className="value code row middle">
-                                            <div className="margin-r-t">
-                                                <a onClick={() => {
-                                                    window.location.href =
-                                                        // eslint-disable-next-line max-len
-                                                        `/${network}/search/${governorAddress}`;
-                                                }}
-                                                >
-                                                    {governorAddress}
-                                                </a>
-                                            </div>
-                                            <CopyButton
-                                                copy={this.state.resolvedIdentity.document.meta.governorAddress}
-                                            />
-                                        </div>
-                                    </div>
+                                </div>
 
-                                    <div className="margin-b-s">
-                                        <div className="label">State Controller</div>
-                                        <div className="value code row middle">
-                                            <div className="margin-r-t">
-                                                <a onClick={() => {
-                                                    // eslint-disable-next-line max-len
-                                                    window.location.href = `/${network}/search/${stateControllerAddress}`;
-                                                }}
-                                                >
-                                                    {stateControllerAddress}
-                                                </a>
-                                            </div>
-                                            <CopyButton
-                                                copy={stateControllerAddress}
-                                            />
+                                <div className="margin-b-s">
+                                    <div className="label">State Controller</div>
+                                    <div className="value code row middle">
+                                        <div className="margin-r-t">
+                                            <a onClick={() => {
+                                                // eslint-disable-next-line max-len
+                                                window.location.href = `/${network}/search/${stateControllerAddress}`;
+                                            }}
+                                            >
+                                                {stateControllerAddress}
+                                            </a>
                                         </div>
+                                        <CopyButton
+                                            copy={stateControllerAddress}
+                                        />
                                     </div>
-                                </Fragment>
+                                </div>
+
+                                <div className="margin-b-s">
+                                    <div className="label">Linked Domains</div>
+                                    <IdentityDomainResolver verifiedDomains={this.state.verifiedDomains} />
+                                </div>
+
+                            </Fragment>
+                        )}
+
+                <div className="margin-b-s">
+                    <h3 className="label">DID Document</h3>
+
+
+                    {!this.state.resolvedIdentity && !this.state.errorMessage && (
+                    <Fragment>
+                        <h3 className="margin-r-s">Resolving DID ...</h3>
+                        <Spinner />
+                    </Fragment>
+                )}
+
+                    {this.state.errorMessage && (
+                    <div className="identity-json-container did-error">
+                        <p className="margin-b-t">ಠ_ಠ </p>
+                        <p className="">{this.state.errorMessage}</p>
+                    </div>
+                 )}
+
+                    {this.state.resolvedIdentity && !this.state.errorMessage && (
+                    <div
+                        className="
+                            json-wraper-stardust-identity
+                            card--value
+                            card--value-textarea
+                            card--value-textarea__json
+                        "
+                    >
+                        <JsonViewer
+                            json={JSON.stringify(
+                                this.state.resolvedIdentity?.document?.doc,
+                                null,
+                                4
                             )}
+                        />
                     </div>
+                )}
                 </div>
-                <div className="section w100">
-                    <div className="section--header">
-                        <h3>
-                            DID Document
-                        </h3>
-                    </div>
-                    <div className="section--data">
 
-                        {!this.state.resolvedIdentity && !this.state.errorMessage && (
-                            <React.Fragment>
-                                <h3 className="margin-r-s">Resolving DID ...</h3>
-                                <Spinner />
-                            </React.Fragment>
-                        )}
+                <div className="margin-b-s">
+                    {this.state.resolvedIdentity && !this.state.errorMessage && (
 
-                        {this.state.errorMessage && (
-                            <div className="identity-json-container did-error">
-                                <p className="margin-b-t">ಠ_ಠ </p>
-                                <p className="">{this.state.errorMessage}</p>
-                            </div>
-                        )}
+                        <div>
+                            <h3 className="label">
+                                Metadata
+                            </h3>
 
-                        {this.state.resolvedIdentity && !this.state.errorMessage && (
                             <div
                                 className="
-                                        json-wraper-stardust-identity
-                                        card--value
-                                        card--value-textarea
-                                        card--value-textarea__json
-                                    "
+                                    json-wraper-stardust-identity
+                                    card--value
+                                    card--value-textarea
+                                    card--value-textarea__json
+                                "
                             >
                                 <JsonViewer
                                     json={JSON.stringify(
-                                        this.state.resolvedIdentity?.document?.doc,
-                                        null,
-                                        4
-                                    )}
-                                />
-                            </div>
-                        )}
-                    </div>
-                </div>
-                <div className="row margin-b-s row--tablet-responsive">
-                    {this.state.resolvedIdentity && !this.state.errorMessage && (
-                        <div className="section w100">
-                            <div className="section--header">
-                                <h3>
-                                    Metadata
-                                </h3>
-                            </div>
-                            <div className="section--data">
-                                <div
-                                    className="
-                                        json-wraper-stardust-identity
-                                        card--value
-                                        card--value-textarea
-                                        card--value-textarea__json
-                                    "
-                                >
-                                    <JsonViewer
-                                        json={JSON.stringify(
                                             this.state.resolvedIdentity?.document?.meta,
                                             null,
                                             3
                                         )}
-                                    />
-                                </div>
-
+                                />
                             </div>
+
                         </div>
                     )}
                 </div>
