+++ client/src/app/components/stardust/address/section/did/DidSection.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import IdentityStardustResolver from "../../../../identity/IdentityStardustResolverNew";
+import IdentityStardustResolver from "../../../../identity/IdentityStardustResolver";
 
 interface DIDSectionProps {
     /**
